Removed dockerfile in string
